move both frontend and backend in same repo + create repository for user table
