+++ backend/src/main/java/com/AlexBurada/Barbershop/controller/BarbershopController.java
@@ -1,0 +1,4 @@
+package com.AlexBurada.Barbershop.controller;
+
+public class BarbershopController {
+}
