+++ backend/src/main/java/com/AlexBurada/Barbershop/controller/UserController.java
@@ -1,0 +1,31 @@
+package com.AlexBurada.Barbershop.controller;
+
+import com.AlexBurada.Barbershop.dto.User;
+import com.AlexBurada.Barbershop.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/users")
+public class UserController {
+
+    @Autowired
+    private UserService service;
+
+    @GetMapping("/")
+    public List<User> getUsers() {
+        return service.getUsers();
+    }
+
+    @GetMapping("/{userId}")
+    public User getUserById(@PathVariable int userId) {
+        return service.getUserById(userId);
+    }
+
+
+}
