+++ backend/src/main/java/com/AlexBurada/Barbershop/dto/Barbershop.java
@@ -4,17 +4,20 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.springframework.stereotype.Component;
 
 @Entity
+@Component
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class UserDTO {
+@Table(name = "barbershop")
+public class Barbershop {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     private String name;
-    private String email;
+    private String city;
+    private String street;
     private String phone;
-    private byte[] profilePicture;
 }
