+++ backend/src/main/java/com/AlexBurada/Barbershop/repository/UserRepo.java
@@ -1,0 +1,10 @@
+package com.AlexBurada.Barbershop.repository;
+
+import com.AlexBurada.Barbershop.dto.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepo extends JpaRepository<User, Integer> {
+
+}
