+++ backend/src/main/resources/application.properties
@@ -1,0 +1,16 @@
+spring.application.name=Barbershop
+
+# This will create the database IF NOT exists
+spring.datasource.url=jdbc:mysql://localhost:3306/barbershop_db?createDatabaseIfNotExist=true
+spring.datasource.username=root
+spring.datasource.password=
+
+#looks for the @Entity classes and updates them and their relations - declare props in this order
+spring.jpa.hibernate.ddl-auto=update
+
+#delay the execution of data.sql file that Hibernate can create tables before initialization
+spring.jpa.defer-datasource-initialization=true
+spring.sql.init.mode=always
+
+#prints the SQL queries that is generated by Spring
+spring.jpa.show-sql=true
