+++ backend/src/main/resources/data.sql
@@ -1,0 +1,19 @@
+INSERT INTO `user` (name, email, phone) VALUES
+                                        ('Alex Burada', 'alex.burada@gmail.com', '0721234567'),
+                                        ('Maria Popescu', 'maria.popescu@yahoo.com', '0734567890'),
+                                        ('Ion Georgescu', 'ion.georgescu@gmail.com', '0745678901'),
+                                        ('Ana Marinescu', 'ana.marinescu@outlook.com', '0756789012'),
+                                        ('Mihai Ionescu', 'mihai.ionescu@gmail.com', '0767890123'),
+                                        ('Elena Stoica', 'elena.stoica@yahoo.com', '0778901234'),
+                                        ('Cristian Radu', 'cristian.radu@gmail.com', '0789012345'),
+                                        ('Diana Popa', 'diana.popa@hotmail.com', '0790123456');
+
+INSERT INTO barbershop (name, city, street, phone) VALUES
+                                       ('Classic Cuts', 'Bucharest', 'Calea Victoriei 15', '0212345678'),
+                                       ('Modern Style Barbershop', 'Cluj-Napoca', 'Strada Memorandumului 28', '0264123456'),
+                                       ('Gentlemen''s Corner', 'Timisoara', 'Bulevardul Revolutiei 7', '0256789012'),
+                                       ('The Barber Shop', 'Iasi', 'Strada Stefan cel Mare 45', '0232456789'),
+                                       ('Elite Cuts', 'Constanta', 'Bulevardul Mamaia 103', '0241567890'),
+                                       ('Royal Barbershop', 'Brasov', 'Strada Republicii 22', '0268234567'),
+                                       ('Urban Cuts', 'Craiova', 'Calea Unirii 89', '0251345678'),
+                                       ('Premium Barbers', 'Galati', 'Strada Domneasca 12', '0236678901');
