--- src/main/java/com/AlexBurada/Barbershop/controller/UserController.java
@@ -1,11 +1,0 @@
-package com.AlexBurada.Barbershop.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/users")
-public class UserController {
-
-
-}
