--- src/main/resources/application.properties
@@ -1,12 +1,0 @@
-spring.application.name=Barbershop
-
-# This will create the database IF NOT exists
-spring.datasource.url=jdbc:mysql://localhost:3306/barbershop_db?createDatabaseIfNotExist=true
-spring.datasource.username=root
-spring.datasource.password=
-
-#looks for the @Entity classes and updates them and their relations
-spring.jpa.hibernate.ddl-auto=update
-
-#prints the actually SQL code that is generated by Spring
-spring.jpa.show-sql=true
